+++ src/main/java/org/spacehq/mc/protocol/data/game/values/world/map/MapPlayer.java
@@ -2,31 +2,31 @@ package org.spacehq.mc.protocol.data.game.values.world.map;
 
 public class MapPlayer {
 
-	private byte iconSize;
-	private byte iconRotation;
-	private byte centerX;
-	private byte centerZ;
+	private int iconSize;
+	private int iconRotation;
+	private int centerX;
+	private int centerZ;
 
 	public MapPlayer(int iconSize, int iconRotation, int centerX, int centerZ) {
-		this.iconSize = (byte) iconSize;
-		this.iconRotation = (byte) iconRotation;
-		this.centerX = (byte) centerX;
-		this.centerZ = (byte) centerZ;
+		this.iconSize = iconSize;
+		this.iconRotation = iconRotation;
+		this.centerX = centerX;
+		this.centerZ = centerZ;
 	}
 
-	public byte getIconSize() {
+	public int getIconSize() {
 		return this.iconSize;
 	}
 
-	public byte getIconRotation() {
+	public int getIconRotation() {
 		return this.iconRotation;
 	}
 
-	public byte getCenterX() {
+	public int getCenterX() {
 		return this.centerX;
 	}
 
-	public byte getCenterZ() {
+	public int getCenterZ() {
 		return this.centerZ;
 	}
