+++ src/main/java/org/spacehq/mc/protocol/data/game/values/world/map/MapScale.java
@@ -2,13 +2,13 @@ package org.spacehq.mc.protocol.data.game.values.world.map;
 
 public class MapScale implements MapData {
 
-	private byte scale;
+	private int scale;
 
 	public MapScale(int scale) {
-		this.scale = (byte) scale;
+		this.scale = scale;
 	}
 
-	public byte getScale() {
+	public int getScale() {
 		return this.scale;
 	}
